+++ spec/dummy_app/config/initializers/secret_token.rb
@@ -1,0 +1,7 @@
+# Be sure to restart your server when you modify this file.
+
+# Your secret key for verifying the integrity of signed cookies.
+# If you change this key, all old signed cookies will become invalid!
+# Make sure the secret is at least 30 characters and all random,
+# no regular words or you'll be exposed to dictionary attacks.
+DummyApp::Application.config.secret_token = 'c4d4c0e87dfb5e0edaf9d5b0ae8012a687171589f7bfc56cae1423cefa02f02cd2aa1e12c1a645d4ee32c3a8040a8d20d95a02a8c8cc7143475836ba3161abc4'