+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version 0.14.0 (unreleased)
 
+* Add a `dojo_config` helper to simplify configuration e.g.:
+
+    <%= dojo_config :base_url => "/assets/dojo"
+
 # Version 0.13.0
 
 *No changes*